requirements/base.txt
pydantic
sentry_sdk
sqlalchemy
typer
passlib
pyjwt
python_dotenv
psutil
rich